% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{incompleteBeta}
\alias{incompleteBeta}
\title{Incomplete beta function}
\usage{
incompleteBeta(x, a, b)
}
\arguments{
\item{x}{Value between 0 and 1 at which to evaluate the incomplete beta
function.}

\item{a}{non-negative numeric}

\item{b}{non-negative numeric}
}
\value{
Returns the value of the incomplete beta function at x.
}
\description{
Computes the incomplete beta function.
}
\details{
This is defined as the integral from 0 to \code{x} of the function
\code{t^(a-1)(1-t)^(b-1)}.
}

