% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{singleWishart}
\alias{doubleWishart}
\alias{singleWishart}
\title{Distribution of the largest root}
\usage{
singleWishart(x, n_min, n_max, mprec = TRUE)

doubleWishart(x, s, m, n, mprec = TRUE)
}
\arguments{
\item{x}{Vector of numeric values at which to compute the CDF.}

\item{n_min, n_max}{Parameters of the single Wishart setting. See details.}

\item{mprec}{Logical. Should we perform high precision computations?}

\item{s, m, n}{Parameters of the double Wishart setting. See Details.}
}
\value{
Returns the value of the CDF at \code{x}.
}
\description{
Computes the cumulative distribution function of the largest root in the single and double Wishart setting.
}
\examples{
x1 <- seq(0, 30, length.out = 50)
y1 <- singleWishart(x1, 5, 10, mprec = FALSE)
plot(x1, y1, type='l')

x2 <- seq(0, 1, length.out = 50)
y2 <- doubleWishart(x2, 10, -0.5, 100, mprec = FALSE)
plot(x2, y2, type='l')
}

