% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doubleWishart}
\alias{doubleWishart}
\alias{singleWishart}
\title{Distribution of the largest root}
\usage{
doubleWishart(x, s, m, n, mprec = TRUE)

singleWishart(x, n_min, n_max, mprec = TRUE)
}
\arguments{
\item{x}{Value at which to compute the CDF.}

\item{s, m, n}{Parameters of the double Wishart setting. See Details.}

\item{mprec}{Logical. Should we use gmp for high precision computations?}

\item{n_min, n_max}{Parameters of the single Wishart setting. See details.}
}
\value{
Returns the value of the CDF at \code{x}.
}
\description{
Computes the cumulative distribution function of the largest root in the single and double Wishart setting.
}

